+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Fixed
+- fixed phone number validation on block-based Checkout page
+
 ## [1.20.4]
 ### Fixed
 - fixed so that the shipping method class would be loaded earlier